+++ apps/docs/stories/Skeleton.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
-import {Skeleton} from "@breadcrumbs/ui";
+import {Skeleton} from "breadcrumbs-ui";
 
 const meta = {
   title: "Components/Skeleton",