+++ packages/breadcrumbs-theme/src/index.tsx
@@ -1,2 +1,3 @@
-export { MediaSizes, mediaQueries } from "./mediaQueries";
-export { colorSafelist, softLight, sand } from "./colors";
+export { MediaSizes, mediaQueries } from './mediaQueries'
+export { colorSafelist, softLight, sand, allThemes } from './colors'
+export type { Colors } from './types'