+++ packages/breadcrumbs-ui/src/Button/types.ts
@@ -1,6 +1,6 @@
-import { Icons } from "@breadcrumbs/icons/src/types";
-import { Colors } from "@breadcrumbs/theme/src/types";
-import { WithButton } from "@breadcrumbs/types/src";
+import { Icons } from "breadcrumbs-icons";
+import { Colors } from "breadcrumbs-theme";
+import { WithButton } from "breadcrumbs-types/src";
 
 export type Sizes = "sm" | "md" | "lg" | "full";
 export type Variants = "primary" | "secondary";